requirements.txt
langchain
langchain_community
langchain-google-genai
pypdf
pdfplumber
faiss-cpu
sentence-transformers
torch
fastapi
uvicorn
python-dotenv